Cancel previous timeout instead of ignoring events
This fixes the IE issue reported to AngularJS at angular/angular.js#16519

Instead of ignoring key events, we cancel the previous deferred timeout and declare a new one each time. 

This fixes the problem that references to `input` and `origValue` inside the deferred function do not point to the most recent key event occurrence.